requirements.txt
# -----------------------------
# Web app
# -----------------------------
flask
flask_sqlalchemy
wtforms
SQLAlchemy
werkzeug
tornado

# -----------------------------
# Scheduler / Task queue
# -----------------------------
apscheduler
celery
kombu

# -----------------------------
# Data & Forecasting
# -----------------------------
yfinance
statsmodels
scikit-learn

# -----------------------------
# Deep Learning
# -----------------------------
transformers
tf-keras

# -----------------------------
# External API / LLM
# -----------------------------
# ollama

# -----------------------------
# Spark
# -----------------------------
pyspark==3.5.1


psycopg2-binary


pandas 
numpy 
tensorflow 
requests 
pytz
flask-migrate




apache-airflow==2.9.2
apache-airflow-providers-apache-spark
pandas
yfinance
statsmodels
scikit-learn
prophet==1.1.5
matplotlib
tensorflow==2.17.0
